# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Books In Budget API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Book {
  id: number;
  title: string;
  author: string;
  description?: string;
  category: string;
  price: number;
  originalPrice: number;
  imageUrl: string;
  rating: number;
  reviewCount: number;
  featured: boolean;
  trending: boolean;
  stockQuantity: number;
  discountPercent: number;
  createdAt?: string;
  updatedAt?: string;
}

export interface BookListResponse {
  books: Book[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CreateBookBody {
  title: string;
  author: string;
  description?: string;
  category: string;
  price: number;
  originalPrice: number;
  imageUrl: string;
  rating?: number;
  reviewCount?: number;
  featured?: boolean;
  trending?: boolean;
  stockQuantity?: number;
}

export interface UpdateBookBody {
  title?: string;
  author?: string;
  description?: string;
  category?: string;
  price?: number;
  originalPrice?: number;
  imageUrl?: string;
  rating?: number;
  reviewCount?: number;
  featured?: boolean;
  trending?: boolean;
  stockQuantity?: number;
}

export interface Category {
  name: string;
  bookCount: number;
  slug: string;
}

export interface BookStats {
  totalBooks: number;
  totalCategories: number;
  avgPrice: number;
  booksUnder199: number;
  trendingCount: number;
  featuredCount: number;
}

export interface InventoryStats {
  totalBooks: number;
  totalStock: number;
  lowStockCount: number;
  outOfStockCount: number;
  totalCategories: number;
  recentlyAdded: Book[];
}

export interface DeleteResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type ListBooksParams = {
  search?: string;
  category?: string;
  minPrice?: number;
  maxPrice?: number;
  featured?: boolean;
  trending?: boolean;
  sortBy?: ListBooksSortBy;
  page?: number;
  limit?: number;
};

export type ListBooksSortBy =
  (typeof ListBooksSortBy)[keyof typeof ListBooksSortBy];

export const ListBooksSortBy = {
  price_asc: "price_asc",
  price_desc: "price_desc",
  rating: "rating",
  title: "title",
} as const;

export type GetTrendingBooksParams = {
  limit?: number;
};

export type GetFeaturedBooksParams = {
  limit?: number;
};

export type GetBudgetBooksParams = {
  maxPrice?: number;
  limit?: number;
};

export type GetRelatedBooksParams = {
  limit?: number;
};